@@ -1,11 +1,11 @@
 name: flutter_siri_suggestions
 description: Flutter plugin for exposure on Siri Suggestions also awake specific function by Siri.
-version: 1.0.2+hotfix1
+version: 1.0.3-nullsafety.0
 author: Chul-won Kim <riky@myriky.net>
 homepage: https://github.com/myriky/flutter_siri_suggestions
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
